requirements.txt
spacy
from spacy.matcher.matcher import Matcher
python-telegram-bot
telegram.ext import Application, CommandHandler, MessageHandler, filters, ConversationHandler
random